# perf: implement performance optimization framework and import efficiency improvements (#57)