--- a/PERFORMANCE.md
+++ b/PERFORMANCE.md
@@ -1,0 +1,40 @@
+# Performance Baseline and Optimization Tracking
+
+## Baseline Measurements (Before Optimization)
+
+### Flake Operations
+- **Flake Show**: 1.182s (real), 0.768s (user), 0.193s (sys)
+- **Flake Check**: 1m22.578s (real), 51.831s (user), 8.046s (sys)
+
+### Host Configuration Evaluation
+- **Thor Evaluation**: 20.550s (real), 14.544s (user), 1.479s (sys)
+
+### Analysis Notes
+- Flake check is the slowest operation at ~82 seconds
+- Individual host evaluation (Thor) takes ~20 seconds
+- Flake show is quite fast at ~1.2 seconds
+- The configuration has grown to 30+ services across 4 hosts
+
+## Optimization Targets
+- [ ] 20% improvement in flake evaluation time (target: <16s for thor)
+- [ ] 15% improvement in flake check time (target: <70s)
+- [ ] Maintain functionality with no regressions
+
+## Optimization History
+
+### Phase 1: Performance Framework (Completed)
+- Added performance measurement targets to Justfile
+- Established baseline measurements
+- Created tracking documentation
+
+### Phase 2: Import Optimization (Pending)
+- Analyze service imports for conditional loading
+- Implement lazy evaluation patterns
+
+### Phase 3: Evaluation Optimization (Pending)
+- Optimize large service modules
+- Improve secrets handling
+
+### Phase 4: Build Optimization (Pending)
+- Configure build caching
+- Optimize flake inputs