--- a/modules/nixos/services/downloads/bazarr.nix
+++ b/modules/nixos/services/downloads/bazarr.nix
@@ -1,11 +1,11 @@
 {
   config,
   lib,
+  nilfheim,
   ...
 }:
 with lib; let
   cfg = config.services.bazarr;
-  constants = import ../../../../lib/constants.nix;
 in {
   options.services.bazarr = {
     url = mkOption {
@@ -56,7 +56,7 @@ in {
       nginx.virtualHosts."${cfg.url}" = {
         locations."/" = {
           proxyPass = "http://127.0.0.1:${toString cfg.listenPort}";
-          inherit (constants.nginxDefaults) proxyWebsockets;
+          inherit (nilfheim.constants.nginxDefaults) proxyWebsockets;
         };
       };
     };